requirements.txt
youtube-transcript-api
langchain
langchain-community
langchain-core
langchain-openai
langchain-google-genai
langchain-huggingface
langchain-qdrant
langchain-experimental
qdrant-client
python-dotenv
fastembed
sentence-transformers
numpy
scikit-learn
transformers
huggingface-hub
scrapetube
streamlit